+++ lib/src/packages/packages.dart
@@ -1,0 +1,4 @@
+export 'localization_package.dart';
+export 'getx_package.dart';
+export 'flutter_intl_package.dart';
+export 'package_factory.dart';